[feat] 관리자 워크스페이스 리스트 조회 시 현재 사용량 float 형태로 추가

+++ module-k8s/src/main/java/com/xiilab/modulek8s/resource_quota/service/ResourceQuotaServiceImpl.java
@@ -63,8 +63,8 @@ public class ResourceQuotaServiceImpl implements ResourceQuotaService {
 			.limitCPU((int)resourceQuotas.getLimitCPU())
 			.limitMEM((int)resourceQuotas.getLimitMEM())
 			.limitGPU((int)resourceQuotas.getLimitGPU())
-			.useCPU((int)resourceQuotas.getUseCPU())
-			.useMEM((int)resourceQuotas.getUseMEM())
+			.useCPU(resourceQuotas.getUseCPU())
+			.useMEM(resourceQuotas.getUseMEM())
 			.useGPU((int)resourceQuotas.getUseGPU())
 			.build()).collect(Collectors.toList());
 	}
